refactor(capabilities): lifecycle as a bridged, required LifecycleCapability the chassis drives
Replace the native generic lifecycle driver with a non-generic bridged capability so a wasm guest can name aether.lifecycle. The graph becomes plain data, the old aether-substrate lifecycle module is deleted, the advance-timeout knob record relocates to the substrate scheduler (its sole consumer), and the three full-stack chassis repoint to the new cap.

Closes iamacoffeepot/aether#1380.
